catchup

+++ AntiVirus/SAV_SimpleAntiVirus/SimpleAntivirus/GUI/Views/Pages/IntegrityPage.xaml.cs
@@ -1,24 +1,10 @@
 ﻿using SimpleAntivirus.GUI.Services;
-using SimpleAntivirus.GUI.ViewModels.Pages;
-using SimpleAntivirus.IntegrityModule.ControlClasses;
 using Microsoft.Win32;
 using SimpleAntivirus.ViewModels.Pages;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Animation;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+using SimpleAntivirus.AntiTampering;
 namespace SimpleAntivirus.GUI.Views.Pages
 {
     /// <summary>
@@ -31,13 +17,23 @@ namespace SimpleAntivirus.GUI.Views.Pages
         private bool _adding;
         public IntegrityPage(IntegrityViewModel integViewModel)
         {
+
             ViewModel = integViewModel;
             DataContext = integViewModel;
             InitializeComponent();
             _adding = false;
         }
 
-        private void Page_Loaded(object sender, RoutedEventArgs e)
+        // Whether to enable or disable buttons to prevent multiple operations at once.
+        private void EnableButton(bool enabler)
+        {
+            IntegrityScanButton.IsEnabled = enabler;
+            AddFile.IsEnabled = enabler;
+            AddFolder.IsEnabled = enabler;
+            DeleteButton.IsEnabled = enabler;
+        }
+
+        private async void Page_Loaded(object sender, RoutedEventArgs e)
         {
             // What to load...
             UpdateEntries();
@@ -56,10 +52,20 @@ namespace SimpleAntivirus.GUI.Views.Pages
 
         private void Delete_Click(object sender, RoutedEventArgs e)
         {
+            EnableButton(false);
             // 0 > no selection, 1> no item found, 2 -> item deleted
-            int resultInt = ViewModel.DeleteItem();
-            DisplayResultDelete(resultInt);
-            ViewModel.PathSelected = null;
+
+            System.Windows.MessageBoxResult choice = System.Windows.MessageBox.Show("Are you sure you want to delete the selected integrity entries?", "Simple Antivirus", System.Windows.MessageBoxButton.OKCancel, MessageBoxImage.Warning);
+            if (choice == System.Windows.MessageBoxResult.OK)
+            {
+                int resultInt = ViewModel.DeleteItem();
+                DisplayResultDelete(resultInt);
+                ViewModel.PathSelected = null;
+                EnableButton(true);
+                return;
+            }
+            DisplayResultDelete(4);
+            EnableButton(true);
         }
 
         // Handle message box for addition info
@@ -68,13 +74,15 @@ namespace SimpleAntivirus.GUI.Views.Pages
             UpdateEntries();
             if (success)
             {
-                MessageBox.Show("Data added to database", "Integrity Add Status", MessageBoxButton.OK, MessageBoxImage.Information);
+                MessageBox.Show("Integrity Entry Added: File(s) successfully added to integrity checking list.", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Information);
             }
             else
             {
-                MessageBox.Show("Data failed to be added to database", "Integrity Add Failure", MessageBoxButton.OK, MessageBoxImage.Error);
+                MessageBox.Show("Integrity Add Failure: File(s) failed to be added.\n Most likely cause is that a file was protected / in use, preventing the program from reading its contents", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
+
+
 
         // Handle message box for addition info
         private void DisplayResultDelete(int returnId)
@@ -83,16 +91,19 @@ namespace SimpleAntivirus.GUI.Views.Pages
             switch (returnId)
             {
                 case 0:
-                    MessageBox.Show("No Item Selected", "Selection Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                    MessageBox.Show("Integrity Entry Delete Failed: No Item Selected", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Error);
                     break;
                 case 1:
-                    MessageBox.Show("Data could not be found in Database", "Incorrect Entry", MessageBoxButton.OK, MessageBoxImage.Error);
+                    MessageBox.Show("Integrity Entry Delete Failed: Entry could not be found in database", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Error);
                     break;
                 case 2:
-                    MessageBox.Show("Data removed successfully", "Data Item Deleted", MessageBoxButton.OK, MessageBoxImage.Information);
+                    MessageBox.Show("Integrity Entries removed successfully!", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Information);
                     break;
                 case 3:
-                    MessageBox.Show("Some data was removed successfully", "Some Data Item Deleted", MessageBoxButton.OK, MessageBoxImage.Information);
+                    MessageBox.Show("Integrity Entries Partially Deleted: Some entries were removed successfully.", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Information);
+                    break;
+                case 4:
+                    System.Windows.MessageBox.Show("Delete operation cancelled. The selected files will continue to be checked for integrity.", "Simple Antivirus", System.Windows.MessageBoxButton.OK, MessageBoxImage.Stop);
                     break;
             }
         }
@@ -123,15 +134,22 @@ namespace SimpleAntivirus.GUI.Views.Pages
 
         private async void AddFile_Click(object sender, RoutedEventArgs e)
         {
+            EnableButton(false);
             bool result = false;
             OpenFileDialog fileDialog = new Microsoft.Win32.OpenFileDialog();
             fileDialog.ShowDialog();
             string fileGet = fileDialog.FileName;
-            if (fileGet != "")
+            // Start AntiTampering Implementation
+            if (InputValSan.FilePathCharLimit(fileGet) && InputValSan.FilePathValidation(fileGet))
             {
-                result = await ViewModel.AddIntegrityPath(fileGet);
-                DisplayResultOfAdded(result);
+                fileGet = InputValSan.FilePathSanitisation(fileGet);
+                if (fileGet != "")
+                {
+                    result = await ViewModel.AddIntegrityPath(fileGet);
+                    DisplayResultOfAdded(result);
+                }
             }
+            EnableButton(true);
         }
 
 
@@ -139,19 +157,26 @@ namespace SimpleAntivirus.GUI.Views.Pages
         // Button that opens folder dialog to be sent to ViewModel.
         private async void AddFolder_Click(object sender, RoutedEventArgs e)
         {
+            EnableButton(false);
             bool result = false;
             OpenFolderDialog folderDialog = new Microsoft.Win32.OpenFolderDialog();
             folderDialog.ShowDialog();
             string folderGet = folderDialog.FolderName;
-            // Start load bar
-            // Send to view model the path of folder.
-            if (folderGet != "")
+            // Start AntiTampering Implementation
+            if (InputValSan.FilePathCharLimit(folderGet) && InputValSan.FilePathValidation(folderGet))
             {
-                DisplayLoading(true);
-                result = await ViewModel.AddIntegrityPath(folderGet);
-                DisplayLoading(false);
-                DisplayResultOfAdded(result);
+                folderGet = InputValSan.FilePathSanitisation(folderGet);
+                // Start load bar
+                // Send to view model the path of folder.
+                if (folderGet != "")
+                {
+                    DisplayLoading(true);
+                    result = await ViewModel.AddIntegrityPath(folderGet);
+                    DisplayLoading(false);
+                    DisplayResultOfAdded(result);
+                }
             }
+            EnableButton(true);
         }
 
         // This is triggered when the table is selected.
@@ -160,28 +185,38 @@ namespace SimpleAntivirus.GUI.Views.Pages
             if (DataShow.SelectedItem != null)
             {
                 List<DataRow> selectedItems = DataShow.SelectedItems.Cast<DataRow>().ToList();
+                int allItemCount = DataShow.Items.Count;
                 string infoText = "";
-                List<string> selectedDirectories = new();
-                if (selectedItems.Count() == 1)
+                if (!(allItemCount == selectedItems.Count) || selectedItems.Count == 1)
                 {
-                    infoText = $"Selected: {selectedItems[0].DisplayDirectory}";
+                    ViewModel.AllSelected = false;
+                    List<string> selectedDirectories = new();
+                    if (selectedItems.Count() == 1)
+                    {
+                        infoText = $"Selected: {selectedItems[0].DisplayDirectory}";
+                    }
+                    else
+                    {
+                        infoText = $"Selected: {selectedItems.Count()} Items";
+                    }
+                    foreach (DataRow datarowItem in selectedItems)
+                    {
+                        selectedDirectories.Add(datarowItem.HiddenDirectory);
+                    }
+                    // Remove final comma.
+                    infoText.Remove(infoText.Length - 1, 1);
+                    ViewModel.PathSelected = selectedDirectories;
                 }
                 else
                 {
-                    infoText = $"Selected: {selectedItems.Count()} Items";
+                    infoText = $"All Items Selected ({allItemCount} Items)";
+                    ViewModel.AllSelected = true;
                 }
-                foreach (DataRow datarowItem in selectedItems)
-                {
-                    selectedDirectories.Add(datarowItem.HiddenDirectory);
-                }
-                // Remove final comma.
-                infoText.Remove(infoText.Length - 1, 1);
-                SelectLabel.Content = infoText;
-                ViewModel.PathSelected = selectedDirectories;
+                SelectLabel.Text = infoText;
             }
             else
             {
-                SelectLabel.Content = "None Selected";
+                SelectLabel.Text = "None Selected";
                 ViewModel.PathSelected = null;
             }
         }
@@ -203,24 +238,38 @@ namespace SimpleAntivirus.GUI.Views.Pages
         {
             if (!ViewModel.ScanInUse)
             {
+                EnableButton(false);
                 int result = await ViewModel.Scan();
                 if (result > 0)
                 {
                     ViolationNote.Foreground = new SolidColorBrush(Colors.Red);
-                    ViolationNote.Content = $"Violations Found: {result}";
+                    ViolationNote.Text = $"Violations Found: {result}";
                     ResultsButton.Visibility = Visibility.Visible;
                 }
                 else
                 {
-                    ViolationNote.Foreground = new SolidColorBrush(Colors.White);
-                    ViolationNote.Content = "No Violations Found";
+                    ViolationNote.ClearValue(TextBlock.ForegroundProperty);
+                    ViolationNote.Text = "No Violations Found";
                 }
+                EnableButton(true);
             }
         }
 
         private void See_Results_Click(object sender, RoutedEventArgs e)
         {
             NavigationServiceIntermediary.NavigationService.Navigate(typeof(IntegrityResultsPage));
+        }
+
+        private void CardExpander_Expanded(object sender, RoutedEventArgs e)
+        {
+            if (IntegrityConfigExpansion.IsExpanded)
+            {
+                IntegrityPageWhole.Height = 984;
+            }
+            else
+            {
+                IntegrityPageWhole.Height = 500;
+            }
         }
     }
 }
